# Membr.Web/openapi/schema.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Membr.API | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5299/"
    }
  ],
  "paths": {
...
    "/integrations/members/by-token/{value}": {
      "get": {
        "tags": [
          "Integrations: Members"
        ],
        "summary": "Resolve a scanned token to its member and membership information",
        "operationId": "LookupMemberByToken",
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenLookupDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TokenLookupDto": {
        "required": [
          "memberId",
          "firstName",
          "surname",
          "memberships"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenLookupMembershipDto"
            }
          }
        }
      },
      "TokenLookupMembershipDto": {
        "required": [
          "membershipTypeName",
          "startDate",
          "endDate",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "membershipTypeName": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}